fix lint

+++ pkg/txmgr/txmgr_test.go
@@ -720,7 +720,7 @@ func TestTxm_GetTransactionFee(t *testing.T) {
 		// insert receipt
 		var r txmgr.Receipt
 		r = newEthReceipt(42, utils.NewHash(), attemptD.Hash, 0x1)
-		expFee := uint64(r.Receipt.EffectiveGasPrice.Int64())*uint64(r.Receipt.GasUsed) + uint64(r.Receipt.L1Fee.Int64())
+		expFee := r.Receipt.EffectiveGasPrice.Uint64()*r.Receipt.GasUsed + r.Receipt.L1Fee.Uint64()
 		_, err = txStore.InsertReceipt(ctx, &r.Receipt)
 		require.NoError(t, err)
